# target/classes/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Payment Initiation — Payment Order API
  description: |
    Contract-first API for initiating and inquiring payment orders, aligned conceptually with
    BIAN Payment Initiation / Payment Order. Migrated from legacy SOAP `SubmitPaymentOrder` and
    `GetPaymentOrderStatus` capabilities.

    **Identifiers:** `paymentOrderId` is an opaque string issued by the service (e.g. UUID, a
    business key such as `PO-0001`, or another internal format). The contract does not fix a
    single format — use the value returned by `POST /payment-initiation/payment-orders` when
    calling the GET endpoints.
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: PaymentOrder
    description: Initiation and retrieval of payment orders (BIAN Payment Initiation)

paths:
  /payment-initiation/payment-orders:
    post:
      tags:
        - PaymentOrder
      summary: Initiate a payment order
      description: |
        Creates a new payment order from client instructions. Equivalent conceptual scope to
        legacy `SubmitPaymentOrder`.
      operationId: createPaymentOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentOrderInitiationRequest'
            examples:
              default:
                summary: Sample initiation (aligned with Postman collection)
                value:
                  externalReference: EXT-1
                  debtorAccount:
                    iban: EC12DEBTOR000000000000
                  creditorAccount:
                    iban: EC98CREDITOR00000000000
                  instructedAmount:
                    amount: 150.75
                    currency: USD
                  remittanceInformation: Factura 001-123
                  requestedExecutionDate: '2025-10-31'
      responses:
        '201':
          description: Payment order created
          headers:
            Location:
              description: URI of the created payment order
              schema:
                type: string
                example: http://localhost:8080/payment-initiation/payment-orders/PO-0001
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentOrderCreationResponse'
              examples:
                created:
                  value:
                    paymentOrderId: PO-0001
                    externalReference: EXT-1
                    status: RECEIVED
                    createdAt: '2025-03-22T10:15:30Z'
        '400':
          description: Bad request — syntactic validation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                validation:
                  value:
                    type: https://api.bank/errors/validation
                    title: Validation failed
                    status: 400
                    detail: One or more fields are invalid.
                    instance: /payment-initiation/payment-orders
        '422':
          description: Unprocessable entity — semantic business rule violation
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                businessRule:
                  value:
                    type: https://api.bank/errors/business-rule
                    title: Business rule violation
                    status: 422
                    detail: Debtor and creditor accounts must differ.
                    instance: /payment-initiation/payment-orders
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  /payment-initiation/payment-orders/{paymentOrderId}:
    get:
      tags:
        - PaymentOrder
      summary: Retrieve a payment order
      description: |
        Returns the full persisted snapshot of a payment order. This extends the legacy SOAP
        surface (which only exposed status inquiry) with a read model suitable for clients.
      operationId: getPaymentOrderById
      parameters:
        - $ref: '#/components/parameters/PaymentOrderId'
      responses:
        '200':
          description: Payment order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentOrder'
              examples:
                full:
                  value:
                    paymentOrderId: PO-0001
                    externalReference: EXT-1
                    debtorAccount:
                      iban: EC12DEBTOR000000000000
                    creditorAccount:
                      iban: EC98CREDITOR00000000000
                    instructedAmount:
                      amount: 150.75
                      currency: USD
                    remittanceInformation: Factura 001-123
                    requestedExecutionDate: '2025-10-31'
                    status: RECEIVED
                    createdAt: '2025-03-22T10:15:30Z'
                    statusChangedAt: '2025-03-22T10:15:30Z'
        '400':
          description: Invalid identifier format
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Payment order not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                notFound:
                  value:
                    type: https://api.bank/errors/not-found
                    title: Payment order not found
                    status: 404
                    detail: No payment order exists for the given identifier.
                    instance: /payment-initiation/payment-orders/PO-0001
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  /payment-initiation/payment-orders/{paymentOrderId}/status:
    get:
      tags:
        - PaymentOrder
      summary: Retrieve payment order status
      description: |
        Lightweight status projection, conceptually aligned with legacy `GetPaymentOrderStatus`.
      operationId: getPaymentOrderStatus
      parameters:
        - $ref: '#/components/parameters/PaymentOrderId'
      responses:
        '200':
          description: Current status of the payment order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentOrderStatusView'
              examples:
                status:
                  value:
                    paymentOrderId: PO-0001
                    status: RECEIVED
                    lastUpdated: '2025-03-22T10:15:30Z'
        '400':
          description: Invalid identifier format
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Payment order not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

components:
  parameters:
    PaymentOrderId:
      name: paymentOrderId
      in: path
      required: true
      description: |
        Internal payment order identifier (opaque string). Use the value returned when the order
        was created.
      schema:
        $ref: '#/components/schemas/PaymentOrderIdentifier'

  schemas:
    PaymentOrderIdentifier:
      type: string
      description: |
        Opaque internal identifier for the payment order (e.g. UUID, business key, or legacy key).
        No single format is mandated by this API.
      minLength: 1
      maxLength: 128

    PaymentOrderInitiationRequest:
      type: object
      additionalProperties: false
      required:
        - externalReference
        - debtorAccount
        - creditorAccount
        - instructedAmount
        - requestedExecutionDate
      properties:
        externalReference:
          type: string
          description: Client-supplied correlation reference (legacy `externalId`).
          minLength: 1
          maxLength: 64
          pattern: '^[\x20-\x7E]+$'
        debtorAccount:
          $ref: '#/components/schemas/AccountReference'
        creditorAccount:
          $ref: '#/components/schemas/AccountReference'
        instructedAmount:
          $ref: '#/components/schemas/InstructedAmount'
        remittanceInformation:
          type: string
          description: Unstructured remittance / narrative (optional).
          maxLength: 140
        requestedExecutionDate:
          type: string
          format: date
          description: Requested execution date for the payment (ISO 8601 calendar date).

    AccountReference:
      type: object
      additionalProperties: false
      required:
        - iban
      properties:
        iban:
          type: string
          description: |
            Account identifier for the party. In production this is often a normalized IBAN; the API
            does not enforce IBAN checksum or structure so that tests, mocks, and sandboxes can use
            dummy values (e.g. short labels, non-IBAN strings). Stricter validation may be applied in
            the service layer when required.
          minLength: 1
          maxLength: 64

    InstructedAmount:
      type: object
      additionalProperties: false
      required:
        - amount
        - currency
      properties:
        amount:
          type: number
          format: double
          description: Monetary amount (JSON number; avoid floating-point in domain implementation).
          minimum: 0.01
        currency:
          type: string
          description: ISO 4217 alphabetic currency code.
          minLength: 3
          maxLength: 3
          pattern: '^[A-Z]{3}$'

    PaymentOrderCreationResponse:
      type: object
      additionalProperties: false
      required:
        - paymentOrderId
        - status
        - createdAt
      properties:
        paymentOrderId:
          $ref: '#/components/schemas/PaymentOrderIdentifier'
        externalReference:
          type: string
          description: Echo of the client reference for correlation.
          maxLength: 64
        status:
          $ref: '#/components/schemas/PaymentOrderStatus'
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp (UTC recommended).

    PaymentOrder:
      type: object
      additionalProperties: false
      required:
        - paymentOrderId
        - externalReference
        - debtorAccount
        - creditorAccount
        - instructedAmount
        - requestedExecutionDate
        - status
        - createdAt
        - statusChangedAt
      properties:
        paymentOrderId:
          $ref: '#/components/schemas/PaymentOrderIdentifier'
        externalReference:
          type: string
          maxLength: 64
        debtorAccount:
          $ref: '#/components/schemas/AccountReference'
        creditorAccount:
          $ref: '#/components/schemas/AccountReference'
        instructedAmount:
          $ref: '#/components/schemas/InstructedAmount'
        remittanceInformation:
          type: string
          maxLength: 140
        requestedExecutionDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/PaymentOrderStatus'
        createdAt:
          type: string
          format: date-time
        statusChangedAt:
          type: string
          format: date-time
          description: Last time the business status changed (maps conceptually to legacy `lastUpdate`).

    PaymentOrderStatusView:
      type: object
      additionalProperties: false
      required:
        - paymentOrderId
        - status
        - lastUpdated
      properties:
        paymentOrderId:
          $ref: '#/components/schemas/PaymentOrderIdentifier'
        status:
          $ref: '#/components/schemas/PaymentOrderStatus'
        lastUpdated:
          type: string
          format: date-time
          description: Timestamp of the last status update (legacy `lastUpdate`).

    PaymentOrderStatus:
      type: string
      description: Business lifecycle state of the payment order.
      enum:
        - RECEIVED
        - PENDING
        - ACCEPTED
        - REJECTED
        - SETTLED
        - FAILED

    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details for HTTP APIs (subset suitable for APIs).
      additionalProperties: true
      required:
        - title
        - status
      properties:
        type:
          type: string
          format: uri-reference
          description: A URI reference that identifies the problem type.
        title:
          type: string
          description: Short, human-readable summary of the problem type.
        status:
          type: integer
          format: int32
          description: HTTP status code.
          minimum: 400
          maximum: 599
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          description: URI reference identifying this specific occurrence.
        code:
          type: string
          description: Optional machine-readable application error code (extension).