# target/generated-sources/openapi/src/main/java/com/hiberius/paymentinitiation/generated/api/PaymentOrderApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.10.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.hiberius.paymentinitiation.generated.api;

import com.hiberius.paymentinitiation.generated.model.PaymentOrder;
import com.hiberius.paymentinitiation.generated.model.PaymentOrderCreationResponse;
import com.hiberius.paymentinitiation.generated.model.PaymentOrderInitiationRequest;
import com.hiberius.paymentinitiation.generated.model.PaymentOrderStatusView;
import com.hiberius.paymentinitiation.generated.model.ProblemDetails;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-22T21:01:34.541313700-05:00[America/Guayaquil]", comments = "Generator version: 7.10.0")
@Validated
public interface PaymentOrderApi {

    /**
     * POST /payment-initiation/payment-orders : Initiate a payment order
     * Creates a new payment order from client instructions. Equivalent conceptual scope to legacy &#x60;SubmitPaymentOrder&#x60;. 
     *
     * @param paymentOrderInitiationRequest  (required)
     * @return Payment order created (status code 201)
     *         or Bad request — syntactic validation failed (status code 400)
     *         or Unprocessable entity — semantic business rule violation (status code 422)
     *         or Internal server error (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/payment-initiation/payment-orders",
        produces = { "application/json", "application/problem+json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<PaymentOrderCreationResponse> createPaymentOrder(
         @Valid @RequestBody PaymentOrderInitiationRequest paymentOrderInitiationRequest
    );


    /**
     * GET /payment-initiation/payment-orders/{paymentOrderId} : Retrieve a payment order
     * Returns the full persisted snapshot of a payment order. This extends the legacy SOAP surface (which only exposed status inquiry) with a read model suitable for clients. 
     *
     * @param paymentOrderId Internal payment order identifier (opaque string). Use the value returned when the order was created.  (required)
     * @return Payment order found (status code 200)
     *         or Invalid identifier format (status code 400)
     *         or Payment order not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/payment-initiation/payment-orders/{paymentOrderId}",
        produces = { "application/json", "application/problem+json" }
    )
    
    ResponseEntity<PaymentOrder> getPaymentOrderById(
        @Size(min = 1, max = 128)  @PathVariable("paymentOrderId") String paymentOrderId
    );


    /**
     * GET /payment-initiation/payment-orders/{paymentOrderId}/status : Retrieve payment order status
     * Lightweight status projection, conceptually aligned with legacy &#x60;GetPaymentOrderStatus&#x60;. 
     *
     * @param paymentOrderId Internal payment order identifier (opaque string). Use the value returned when the order was created.  (required)
     * @return Current status of the payment order (status code 200)
     *         or Invalid identifier format (status code 400)
     *         or Payment order not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/payment-initiation/payment-orders/{paymentOrderId}/status",
        produces = { "application/json", "application/problem+json" }
    )
    
    ResponseEntity<PaymentOrderStatusView> getPaymentOrderStatus(
        @Size(min = 1, max = 128)  @PathVariable("paymentOrderId") String paymentOrderId
    );

}